Added a /metrics handler to fetch metrics data
Right now we're writing metrics data to a file. This new handler allows
others to fetch metrics data. Right now the handler only returns the
counts for the last 24 hours as a temporary thing. This should be changed,
depending on what the metrics team needs but definitely to include
measurements from the last full 24 hour measurement period.