Add a sessionID to sequencing header
The sequencing layer needs a sessionID to identify client sessions
across different proxies. This is especially important in the case that
the client sends some data and then the proxy stops working reliably and
wants to resend the data over another proxy or continue their connection
with the bridge without restarting.