Added timeout for acknowledged data
Added a timer to all calls to the SnowflakeConn Write() method. The
timer will close the underlying connection if sent data isn't
acknowledged within a timeout duration. This duration is a constant
currently set at 10 seconds, but can be easily changed. Also added unit
tests to check that the connection closes if the data has not been
acknowledged.