Save and resend unacknowledged bytes
When a client or server make a call to SnowflakeConn's Write() method,
the bytes written will be saved in a buffer until they are acknowledged.
If they are not acknowledged within the given timeframe, the underlying
connection will be closed and SnowflakeConn will wait to be given a new
snowflake. When the new snowflake is set, the unacknowledged bytes from
the previous snowflake will be resent.