Implemented a sequencing layer protocol
This implements a protocol between snowflake clients and snowflake
servers that allows for sequencing and reliability that is persistent
across snowflakes.

The protocol sends data in chunks, accompanied by a header with sequence
and acknowledgement numbers. Data is acknowledged once it is received
and unacknowledged data is resent when a new snowflake is added.