Restructured the sending/reading of sessionID
Throughtout writing this, I've tried to keep the protocol as symmetric
as possible for the client and the server, but there is one key
difference between these two endpoints: the server needs to read the
session ID of new incomming client connections.

This change moves the session ID out of every single packet and to the
very first messages sent in a new connection from the client. The server
side then needs to call ReadSessionID before starting to Read data from
the connection in the usual way.

We might want to end up modifying these function calls to be more
different for clients and servers in the end.