Have broker inform proxies of updates
This is targeted specifically at the standalone go proxies. Proxies
report their version to the broker and the broker will include an
additional field in their response to indicate whether the proxies
should update. If there is a newer version available, a log message will
br printed after every poll.